+++ WebPlayGames/PHP/Clases.php
@@ -1,3 +1,2 @@
 <?php
 require_once 'Conexion.php';
-require_once 'Consultas.php';
